+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @search-docs/cli
 
+## 1.0.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @search-docs/types@1.0.6
+  - @search-docs/server@1.0.6
+  - @search-docs/client@1.0.6
+
 ## 1.0.5
 
 ### Patch Changes