+++ packages/client/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @search-docs/client
 
+## 1.0.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @search-docs/types@1.0.6
+
 ## 1.0.5
 
 ### Patch Changes