+++ packages/storage/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @search-docs/storage
 
+## 1.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @search-docs/types@1.0.6
+
 ## 1.0.2
 
 ### Patch Changes